web/satellite: Update login area to use latest production code
This is another step in removing code from our branding repository. The
code between our three prod satellites was almost identical - this
change just uses the html/scss for the prod satellites, and generalizes
a small portion of the html so that the satellite selection dropdown is
based on configuration rather than hardcoded.

There are significant differences in the code for non-prod satellites
like saltlake, eu-north, and us2. We can address this in a future
change.

Change-Id: I72a79f5b60670abb6340c62132b35ba3884d0750