web/storagenode: add online score block for selected satellite
Online score is shown on the dashboard for each satellite when all the satellites details are displayed.
When a user selects a single satellite, there is no online score block on the dashboard for the selected satellite.
This change adds an online score block to the suspension and audit area for case when a specific satellite is selected

Resolves https://github.com/storj/storj/issues/4180

Change-Id: I6065b64b9d9a7def946c1cc521494e1927799cfa