testsuite/api/graphql:Implements Introspection Test
Rework of introspection tests to use embeded structs for Client/Server
implementation of test. This improves the code and helps to remove
panics on errors. Logging was added with zap and errs libraries.
endpoints.go was removed as we no longer need a main funciton to call
these tests, instead we are using the _test.go/packagename_test
nomenclature to be consistent with other tests in the repo.