create go-tooling.yml
implement go tooling on separate workflow. By breaking these into individual workflows we are able to run more tests in parallel. In the next few iterations we will need to cache the built binaries so that each of these sub-workflows doesn't have to fetch the repo and build the needed binaries.